chore: moved cookie validator to authcontroller
